autoscaler/requirements.txt
google-cloud-pubsub==2.18.4
docker==6.1.3
prometheus-client==0.19.0
flask==3.0.0

